fix(cuda_polar_voxel_outlier_filter): correct meaningless cast

+++ sensing/autoware_cuda_pointcloud_preprocessor/src/cuda_outlier_filter/cuda_polar_voxel_outlier_filter.cu
@@ -143,13 +143,13 @@ __device__ void atomic_add_size_t(size_t * addr, size_t val)
 __device__ [[nodiscard]] inline bool meets_primary_threshold(
   const size_t & count, const int & threshold)
 {
-  return count >= static_cast<int>(threshold);
+  return count >= static_cast<size_t>(threshold);
 }
 
 __device__ [[nodiscard]] inline bool meets_secondary_threshold(
   const size_t & count, const int & threshold)
 {
-  return count <= static_cast<int>(threshold);
+  return count <= static_cast<size_t>(threshold);
 }
 
 __device__ [[nodiscard]] inline bool meets_intensity_threshold(
